+++ es/api/connect-dumb.js
@@ -1,7 +1,0 @@
-
-import registerDumb from './register-dumb';
-
-export default function (connectSpec, ccClassKey) {
-  return registerDumb({ connect: connectSpec }, ccClassKey);
-}
-