+++ es/api/connect.js
@@ -1,7 +1,0 @@
-
-import register from './register';
-
-export default function (connectSpec, ccClassKey) {
-  return register({ connect: connectSpec }, ccClassKey);
-}
-