+++ es/api/dispatch.js
@@ -1,5 +1,0 @@
-import dispatch from '../core/base/dispatch';
-
-export default function (action, payLoadWhenActionIsString, rkOrOptions, delay, extra) {
-  return dispatch(action, payLoadWhenActionIsString, rkOrOptions, delay, extra);
-}