+++ es/api/get-computed.js
@@ -1,8 +1,0 @@
-import ccContext from '../cc-context';
-
-const _computedValue = ccContext.computed._computedValue;
-
-export default module => {
-  if (module) return _computedValue[module];
-  else return _computedValue;
-}