+++ es/api/get-connected-state.js
@@ -1,8 +1,0 @@
-import ccContext from '../cc-context';
-
-const ccClassKey_ccClassContext_ = ccContext.ccClassKey_ccClassContext_;
-
-export default ccClassKey => {
-  const ctx = ccClassKey_ccClassContext_[ccClassKey];
-  return ctx.connectedState || {};
-}