+++ es/api/get-global-computed.js
@@ -1,8 +1,0 @@
-import ccContext from '../cc-context';
-import { MODULE_GLOBAL } from '../support/constant';
-
-const _computedValue = ccContext.computed._computedValue;
-
-export default () => {
-  return _computedValue[MODULE_GLOBAL];
-}