+++ es/api/get-refs.js
@@ -1,4 +1,0 @@
-
-import getRefs from '../core/ref/get-refs';
-
-export default getRefs;