+++ es/api/reducer.js
@@ -1,5 +1,0 @@
-import ccContext from '../cc-context';
-
-const _reducerCaller = ccContext.reducer._reducerCaller;
-
-export default _reducerCaller;