+++ es/core/base/catch-cc-error.js
@@ -1,6 +1,0 @@
-import ccContext from '../../cc-context';
-
-export default err=>{
-  if(ccContext.errorHandler)ccContext.errorHandler(err);
-  else throw err;
-}