+++ es/core/state/set-connected-state.js
@@ -1,6 +1,0 @@
-import * as util from '../../support/util'
-
-export default function (connectedState, module, key, value) {
-  const moduleConnState = util.safeGetObjectFromObject(connectedState, module);
-  moduleConnState[key] = value;
-}