+++ es/core/watch/get-watch-spec.js
@@ -1,5 +1,0 @@
-
-export default function (watchFns, immediateWatchKeys = []) {
-  let hasFn = Object.keys(watchFns).length > 0;
-  return { watchFns, immediateWatchKeys, hasFn };
-}