+++ es/support/share.js
@@ -1,9 +1,0 @@
-import ccContext from '../cc-context';
-
-export function strictWarning(err) {
-  if (ccContext.isStrict) {
-    throw err;
-  }
-  console.error(' ------------ CC WARNING ------------');
-  console.error(err);
-}