+++ src/core/watch/watch-key-for-ref.js
@@ -1,12 +1,13 @@
 import pickDepFns from '../base/pick-dep-fns';
 import findDepFnsToExecute from '../base/find-dep-fns-to-execute';
+import { CATE_REF } from '../../support/constant';
 
 export default function (refCtx, stateModule, oldState, committedState, callInfo, isBeforeMount, autoMergeDeltaToCommitted = false) {
 
   if (!refCtx.hasWatchFn) return true;
 
   const deltaCommittedState = Object.assign({}, committedState);
-  const { watchDep, module: refModule, ccUniqueKey } = refCtx;
+  const { watchDep, module: refModule, ccUniqueKey, refComputed, refConnectedComputed } = refCtx;
   const newState = Object.assign({}, oldState, committedState);
 
   const curDepComputedFns = (committedState, isBeforeMount) => pickDepFns(isBeforeMount, 'ref', 'watch', watchDep, stateModule, oldState, committedState, ccUniqueKey);
@@ -14,6 +15,7 @@ export default function (refCtx, stateModule, oldState, committedState, callInfo
   const shouldCurrentRefUpdate = findDepFnsToExecute(
     refCtx, stateModule, refModule, oldState, curDepComputedFns,
     committedState, newState, deltaCommittedState, callInfo, isBeforeMount,
+    'watch', CATE_REF, refComputed, refConnectedComputed,
   );
 
   if (autoMergeDeltaToCommitted) {